Severity Prediction of Knee Osteoarthritis
Knee Osteoarthritis (OA) is a condition that affects the joints, including the knee joint, due to the significant weight-bearing of the body. The disease results in degeneration and rupture of the cartilage elements in the knee joint, which causes severe pain. Unfortunately, the prevalence of OA has been increasing globally, with a 113.25% increase in cases from 1990 to 2019. Currently, more than 350 million people globally suffer from arthritis, and it is estimated that by 2040, about 78 million US adults will have the condition. The diagnosis of OA is primarily carried out by evaluating symptoms and comparing plain radiographs, which can be subjective. However, deep learning techniques, such as Convolution Neural Networks (CNNs), have emerged as a promising solution to medical problems in recent years. Therefore, the goal of this study is to develop and implement a classification system that can aid doctors in reducing their workload and assist radiologists in assessing the severity of the pain accurately. Furthermore, this will enable them to make the  best diagnosis and recommend the most appropriate treatment. One of the methods used in detecting knee OA is transfer learning with fine-tuning, which leverages deep learning techniques to improve the accuracy of diagnosis and classification. By using this method, doctors and radiologists can make informed decisions on the severity of the condition, which can lead to better treatment outcomes for patients.

The following are the goals of the study:
1. The primary objective of this research is to analyze the degree of severity of knee x-ray images.
2. To assist Rheumatologists, in identifying the severity of the disease after the medication, thus reducing the time taken and lengthy treatments.
3. For monitoring and implementing early treatments to sustain good health in patients who pose a concern.
4. To check how accurately, the degree of severity of knee x-ray photographs is estimated by the deep learning algorithm.

The following are the social causes of the study:
1. To assist Rheumatologists, in identifying the severity of the disease after the medication, thus reducing the time taken and lengthy treatments.
2. To reduce the medical and travelling expenses.
3. To overcome the manual mistakes made by the doctor sometimes.
4.  If any person got medication before , so he can check his disease in our webpage rather than consulting any doctor.