Knee Osteoarthritis
severiety Prediction of Knee Osteoarthritis
Osteoarthritis (OA) of the knee is an inflammation that impacts the knee bone, due to the significant weight-bearing of the body. The disease results in degeneration and rupture of the cartilage elements in the knee joint, which causes severe pain. Unfortunately, the prevalence of OA has been increasing globally, with a 113.25% increase in cases from 1990 to 2019. Currently, more than 350 million people globally suffer from arthritis, and it is estimated that by 2040, about 78 million US adults will have this condition. The diagnosis of OA is primarily carried out by evaluating symptoms and comparing plain radiographs, which can be subjective. However, Convolution Neural Networks (CNNs), one of the best deep learning technological advances, are currently attracting interest as a potential remedy for healthcare problems. Consequently, this study's objective aims to design and implement a categorization scheme that can aid doctors in reducing their workload and assist rheumatologists in assessing the severity of the pain accurately. Furthermore, this will enable them to make the best diagnosis and recommend the most appropriate treatment. By using our proposed model i.e., Enhanced MobileNet-v2, rheumatologists can make informed decisions on the severity of the condition, which can lead to better treatment outcomes for the patients. The proposed model Enhanced MobileNet-V2 was trained on the digital knee x-ray image dataset and achieved 91%  accuracy

The following are the goals of the study:
1. The primary objective of this research is to analyze the degree of severity of knee x-ray images.
2. To assist Rheumatologists, in identifying the severity of the disease after the medication, thus reducing the time taken and lengthy treatments.
3. For monitoring and implementing early treatments to sustain good health in patients who pose a concern.
4. To check how accurately, the degree of severity of knee x-ray photographs is estimated by the deep learning algorithm.

The following are the social causes of the study:
1.To assist Rheumatologists, in identifying the severity of the disease after the medication, thus reducing the time taken and lengthy treatments.
2. To reduce the medical and travelling expenses.
3. To overcome the manual mistakes made by the doctor sometimes.
4. If any person got any medication before , he can check his disease severiety in our webpage rather than consulting the doctor again.